Always use inline Medications for MedicationRequests
That is, avoid medicationReference in favor of
medicationCodeableConcept for any generated MedicationRequest.

This is to avoid downstream Cumulus ETL difficulty in processing them.
We could make the ETL smarter, but this is an easier fix for the moment,
since no real-world EHR needs the ability to ingest separate medications
yet.